--- hw06/src/main/java/ru/otus/spring/hw05/domain/Comment.java
@@ -1,25 +1,0 @@
-package ru.otus.spring.hw05.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-
-@Entity
-@Table(name = "comments")
-public class Comment {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "book_id")
-    private Long bookId;
-
-    @Column(name = "comment", nullable = false, unique = false)
-    private String comment;
-}
