--- hw06/src/main/java/ru/otus/spring/hw05/domain/Genre.java
@@ -1,22 +1,0 @@
-package ru.otus.spring.hw05.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-
-@Entity
-@Table(name = "genres")
-public class Genre {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "description", nullable = false, unique = true)
-    private String description;
-}
