--- hw06/src/main/java/ru/otus/spring/hw05/repository/impl/CommentRepositoryJpa.java
@@ -1,53 +1,0 @@
-package ru.otus.spring.hw05.repository.impl;
-
-import org.springframework.stereotype.Repository;
-import ru.otus.spring.hw05.domain.Comment;
-import ru.otus.spring.hw05.repository.CommentRepository;
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.Query;
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public class CommentRepositoryJpa implements CommentRepository {
-
-    @PersistenceContext
-    private EntityManager em;
-
-    @Override
-    public Comment save(Comment comment) {
-        if(comment.getId() == null) {
-            em.persist(comment);
-            return comment;
-        } else {
-            return em.merge(comment);
-        }
-    }
-
-    @Override
-    public void updateCommentById(long id, String comment) {
-        Query query = em.createQuery("update Comment c set c.comment = :comment where c.id = :id");
-        query.setParameter("comment", comment);
-        query.setParameter("id", id);
-        query.executeUpdate();
-    }
-
-    @Override
-    public Optional<Comment> findById(long id) {
-        return Optional.ofNullable(em.find(Comment.class, id));
-    }
-
-    @Override
-    public List<Comment> findAll() {
-        Query query = em.createQuery("select c from Comment c ");
-        return query.getResultList();
-    }
-
-    @Override
-    public void deleteById(Long id) {
-        Query query = em.createQuery("delete from Comment c where c.id = :id");
-        query.setParameter("id", id);
-        query.executeUpdate();
-    }
-}
