--- hw06/src/main/java/ru/otus/spring/hw05/service/CommentService.java
@@ -1,13 +1,0 @@
-package ru.otus.spring.hw05.service;
-
-import ru.otus.spring.hw05.domain.Comment;
-
-import java.util.List;
-
-public interface CommentService {
-    void add(Comment comment);
-    void update(Long id, String comment);
-    Comment getById(long id);
-    List<Comment> getAll();
-    void delete(long id);
-}
