+++ hw06/src/main/java/ru/otus/spring/hw06/Runner.java
@@ -1,0 +1,11 @@
+package ru.otus.spring.hw06;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Runner {
+	public static void main(String[] args) {
+		SpringApplication.run(Runner.class, args);
+	}
+}
