+++ hw06/src/main/java/ru/otus/spring/hw06/domain/Comment.java
@@ -1,0 +1,26 @@
+package ru.otus.spring.hw06.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+
+@Entity
+@Table(name = "comments")
+public class Comment {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne()
+    @JoinColumn(name = "book_id")
+    private Book book;
+
+    @Column(name = "text", nullable = false)
+    private String text;
+}
