+++ hw06/src/main/java/ru/otus/spring/hw06/repository/AuthorRepository.java
@@ -1,0 +1,13 @@
+package ru.otus.spring.hw06.repository;
+
+import ru.otus.spring.hw06.domain.Author;
+import java.util.List;
+import java.util.Optional;
+
+public interface AuthorRepository {
+    Author save(Author author);
+    void updateNameById(long id, String name);
+    Optional<Author> findById(long id);
+    List<Author> findAll();
+    void deleteById(Long id);
+}
