+++ hw06/src/main/java/ru/otus/spring/hw06/repository/BookRepository.java
@@ -1,0 +1,13 @@
+package ru.otus.spring.hw06.repository;
+
+import ru.otus.spring.hw06.domain.Book;
+import java.util.List;
+import java.util.Optional;
+
+public interface BookRepository {
+    Book save(Book genre);
+    void updateTitleById(long id, String title);
+    Optional<Book> findById(long id);
+    List<Book> findAll();
+    void deleteById(Long id);
+}
