+++ hw06/src/main/java/ru/otus/spring/hw06/repository/impl/BookRepositoryJpa.java
@@ -1,0 +1,54 @@
+package ru.otus.spring.hw06.repository.impl;
+
+import org.springframework.stereotype.Repository;
+import ru.otus.spring.hw06.domain.Book;
+import ru.otus.spring.hw06.repository.BookRepository;
+import javax.persistence.*;
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+public class BookRepositoryJpa implements BookRepository {
+
+    @PersistenceContext
+    private EntityManager em;
+
+    @Override
+    public Book save(Book book) {
+        if(book.getId() == null) {
+            em.persist(book);
+            return book;
+        } else {
+            return em.merge(book);
+        }
+    }
+
+    @Override
+    public void updateTitleById(long id, String title) {
+        Query query = em.createQuery("update Book b set b.title = :title where b.id = :id");
+        query.setParameter("title", title);
+        query.setParameter("id", id);
+        query.executeUpdate();
+    }
+
+    @Override
+    public Optional<Book> findById(long id) {
+        return Optional.ofNullable(em.find(Book.class, id));
+    }
+
+    @Override
+    public List<Book> findAll() {
+        EntityGraph<?> bookAuthorGenreEntityGraph = em.getEntityGraph("book-author-genre-entity-graph");
+        TypedQuery<Book> query = em.createQuery("select b from Book b", Book.class);
+        query.setHint("javax.persistence.fetchgraph", bookAuthorGenreEntityGraph);
+        return query.getResultList();
+    }
+
+    @Override
+    public void deleteById(Long id) {
+        Query query = em.createQuery("delete from Book b where b.id = :id");
+        query.setParameter("id", id);
+        query.executeUpdate();
+    }
+
+}
