+++ hw06/src/test/resources/data.sql
@@ -4,9 +4,9 @@ insert into GENRES (id, description) values (1, 'Test Fantasy');
 insert into GENRES (id, description) values (2, 'Test Fantasy second');
 insert into BOOKS (id, title, author_id, genre_id) values (1, 'Test Lord of the Rings', 1, 1);
 insert into BOOKS (id, title, author_id, genre_id) values (2, 'Test Lord of the Rings episode 2', 2, 2);
-insert into COMMENTS (id, book_id, comment) values (1, 1, 'Test comment');
-insert into COMMENTS (id, book_id, comment) values (2, 1, 'Test comment');
-insert into COMMENTS (id, book_id, comment) values (3, 2, 'Test comment');
-insert into COMMENTS (id, book_id, comment) values (4, 2, 'Test comment');
-insert into COMMENTS (id, book_id, comment) values (5, 2, 'Test comment');
+insert into COMMENTS (id, book_id, text) values (1, 1, 'Test comment');
+insert into COMMENTS (id, book_id, text) values (2, 1, 'Test comment');
+insert into COMMENTS (id, book_id, text) values (3, 2, 'Test comment');
+insert into COMMENTS (id, book_id, text) values (4, 2, 'Test comment');
+insert into COMMENTS (id, book_id, text) values (5, 2, 'Test comment');
